# fix `Math.pow_body` for huge powers