Fix `Math.pow_body(a, n)` for huge values of `n`.

Fixes #53881