@KlausC 2024-03-28
> Is my understanding correct that the theory behind this is that it moves the first iteration out of the loop? It's somewhat unclear to me why this makes a difference.

Yes, I took the first execution out of the loop, because I wanted to half the size of `n`, so `n = -n` does not break.
As a side-effect, the inverse of `x` was not required, but of `x * x`, which made the difference for `x = prevfloat(1.0)`.