@oscardssmith 2024-03-28
Overall this looks good. I do want to do a little more testing with random giant comments to make sure this fully fixes the issues.