@KlausC 2024-04-01
The `pow_body` method was face-lifted a little bit. Now it has a accuracy of a few ulps in the complete definition area with no performance regressions. Please see the reported results:

```
function ulps(x, n, p=^)
           a = p(x, n)
           b = big(x) ^ n
           u = abs(a - Float64(b)) / eps(a)
           return isfinite(u) ? Int128(ceil(u)) : -1
 end

function stat_ulps(m)
           M1, M2 = log(floatmin(Float64)), log(floatmax(Float64))
           l1, l2 = sqrt(log2(3)), sqrt(log2(log2(floatmax()) / log2(nextfloat(1.0))))
           D = Dict{Integer,Any}()
           try
             for i = 1:m
               ab = rand() * (M2 - M1) + M1
               lb = (rand() * (l2 - l1) + l1) ^ 2
               b = Int64(floor(2.0 ^ lb)) * rand([-1,1])
               ab /= b
               s = sign(ab)
               la = prevfloat(abs(ab))
               c = expm1(la)
               a = (c - eps(c + 1) / 2) + 1
               a = s < 0 ? inv(a) : a
               if !isfinite(a)
                   println("ab=$ab lb=$lb s=$s b=$b la=$la c=$c a=$a")
               end
               u = ulps(a, b)
               R = get!(D, u) do; [] end
               t = (b, a)
               push!(R, t)
             end
           finally
               ;
           end
           for k in keys(D)
               D[k] = sort(collect(Set(D[k])); by = t-> abs(t[1]) + log2(abs(t[2])+eps()) * eps())
           end
           D
end

 function overview(D::Dict{<:Integer,<:Any})
         s0 = s1 = s2 = 0.0
         m = NaN
         ky = sort!(collect(keys(D)))
         n = sum(length(D[k]) for k in ky); println(n)
          for k in ky
                 l = float(length(D[k]))
                 if s0 + l > n / 2 && isnan(m)
                      m = k - 0.5 + (n/2 - s0) / l
                 end
                 s0 += l
                 s1 += l * k
                 s2 += l * k * k
                  println("$k => $l")
           end
           u = round(s1 / s0; digits = 3)
            v = round(sqrt( s2 / s0 - u * u); digits=3)
            u, m, v = round.((u, m, v); digits=3)
            println("total: n: $s0 - median: $m - mean: $u ± $v ulps")
end

julia> D = @time stat_ulps(10^6);
  6.709275 seconds (9.64 M allocations: 471.625 MiB, 8.45% gc time, 2.49% compilation time)

julia> overview(D)
1000000
0 => 768967.0
1 => 230895.0
2 => 77.0
3 => 39.0
4 => 9.0
5 => 8.0
6 => 4.0
7 => 1.0
total: n: 1.0e6 - median: 0.15 - mean: 0.231 ± 0.423 ulps

julia> @benchmark x^n setup= n, x = rand(D[0])
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  18.704 ns …  6.537 μs  ┊ GC (min … max): 0.00% … 98.42%
 Time  (median):     69.834 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   80.907 ns ± 84.292 ns  ┊ GC (mean ± σ):  1.07% ±  1.38%

   ▂█▂                                                         
  ▃███▇▇▇▆▅▄▅▅▅▅▅▅▅▃▄▄▅▅▄▅▅▄▃▄▃▄▄▄▄▄▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  18.7 ns         Histogram: frequency by time         253 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.

julia> @benchmark x^n setup= n, x = rand(D[1])
BenchmarkTools.Trial: 10000 samples with 997 evaluations.
 Range (min … max):  18.748 ns …  5.083 μs  ┊ GC (min … max): 0.00% … 98.91%
 Time  (median):     72.091 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   83.205 ns ± 74.017 ns  ┊ GC (mean ± σ):  0.81% ±  1.31%

   ▅█▄▃▃▂▁                                                     
  ▄███████▇▅▇██▇▇▇▆▅▅▇▆▅▆▆▆▅▄▄▅▆▅▅▅▄▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  18.7 ns         Histogram: frequency by time         265 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.

julia> @benchmark x^n setup= n, x = rand(D[2])
BenchmarkTools.Trial: 10000 samples with 949 evaluations.
 Range (min … max):   96.888 ns …  5.754 μs  ┊ GC (min … max): 0.00% … 94.29%
 Time  (median):     280.251 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   276.585 ns ± 96.419 ns  ┊ GC (mean ± σ):  0.30% ±  1.29%

                                  █▁▂▃    ▁                     
  ▄▁▂▄▂▄▂▄▄▆▅▂▂▂▂▃▄▃▂▂▄▂▂▅▂▂▄▂▂▃▆▆████▂▅▃▇█▃▂▂▄▅▅▄▅▅▆▆▄▃▂▄▂▂▂▃ ▃
  96.9 ns         Histogram: frequency by time          413 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.

julia> @benchmark x^n setup= n, x = rand(D[3])
BenchmarkTools.Trial: 10000 samples with 852 evaluations.
 Range (min … max):  143.079 ns …  6.419 μs  ┊ GC (min … max): 0.00% … 94.95%
 Time  (median):     304.870 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   311.604 ns ± 83.042 ns  ┊ GC (mean ± σ):  0.20% ±  0.95%

                       ▄▁▁█  ▅▇ ▂   ▁▂                          
  ▅▂▁▂▂▂▁▁▁▁▁▁▁▅▃▁▁▁▁▃▅████▇▇██▄█▅▇▆██▆▂▅▆▂▂▂▂▂▂▂▂▂▂▂▂▂▁▅▃▁▁▁▄ ▃
  143 ns          Histogram: frequency by time          490 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.
```