@KlausC 2024-04-02
This is now ready to be reviewed / merged. Note that it is an essential improvement in accuracy (only 150 ppm more than 1 ULP error) for big exponents, while performance is maintained. (The current implementation has more than 34% in the same sampled argument area)